Make sidebar animation performant.
This commit changes the way the main content is resized, so it doesn't use margin. From an idea by Matías, we now set the destination width _instantly_ when the sidebar is toggled. The position of the editor is then translated to its destination. The result looks pretty good, even though things like sticky, extlink and others aren't smoothly animated. CC: @shaunandrews