Work around the position:fixed & transform issue.
This commit uses an animation to translate the sidebar when it needs to animate. But every animation ends with unsetting the translate, so that the fixed position MCE toolbar can work.

It is a little bit hacky, but it works pretty decently, and certainly feels faster than animating the margin.

Two down-sides:

- if you are scrolled on a long page, and the MCE toolbar is fixed to the top, then toggle the sidebar, the MCE toolbar will disappear and reappear
- there's a slight stutter at the end of each animation, when the translate is unset.

Let's try this, feel it, see if it works well. If we don't like this, the best alternative at this point seems to be not animating the body content at all, simply set it instantly. That might not be as smooth, but at least won't be stuttery.