Build: Add our own custom stable module id, chunk id, and chunk hash generators.
The module ID generator is lifting mostly from webpack-stable-module-id-and-hash. It works quite well, and could be replaced with the new HashedModuleIdPlugin when we move to webpack 2.

The chunk ID generator is based on it, but uses the chunk name instead of the module path as a key into the chunk ID.

The chunk hash generator is based on webpack-chunk-hash, with the addition of the dependant chunks in the hash content. This breaks the hash when the chunk ids of dependent chunks change. With the stable chunk module this shouldn't happen, but in case we drop it, add it for safety.